+++ crates/object-builder/src/buffer_stream/buffer.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use serde_json::Value;
 
 use crate::view_plan::ViewPlan;
-use crate::{ObjectIdPair, RowSource};
+use crate::{sources::RowSource, ObjectIdPair};
 
 /// Because objects are received on the go, and object builder needs to create joins,
 /// these objects need to be tempoirairly stored in some kind of buffer until the last part
