+++ crates/object-builder/src/lib.rs
@@ -1,9 +1,6 @@
 use async_trait::async_trait;
 use bb8::Pool;
-use cdl_dto::{
-    edges::TreeResponse,
-    materialization::{self, FieldType},
-};
+use cdl_dto::{edges::TreeResponse, materialization};
 use communication_utils::{consumer::ConsumerHandler, message::CommunicationMessage};
 use futures::{future::ready, Stream, StreamExt, TryStreamExt};
 use metrics_utils::{self as metrics, counter};
@@ -29,6 +26,7 @@ mod buffer_stream;
 mod view_plan;
 
 mod row_builder;
+mod sources;
 
 #[derive(Clone)]
 pub struct ObjectBuilderImpl {
@@ -123,53 +121,6 @@ mod object_id_pair {
             deserializer.deserialize_str(Helper)
         }
     }
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize)]
-pub enum FieldDefinitionSource {
-    Simple {
-        object: ObjectIdPair,
-        field_name: String,
-        field_type: FieldType,
-    },
-    Computed {
-        computation: ComputationSource,
-        field_type: FieldType,
-    },
-    Array {
-        fields: HashMap<String, FieldDefinitionSource>,
-    },
-}
-
-#[derive(Debug, PartialEq, Deserialize, Clone, Serialize)]
-pub enum ComputationSource {
-    RawValue {
-        value: Value,
-    },
-    FieldValue {
-        object: ObjectIdPair,
-        field_path: String,
-    },
-    Equals {
-        lhs: Box<ComputationSource>,
-        rhs: Box<ComputationSource>,
-    },
-}
-
-#[derive(Debug, PartialEq)]
-pub enum RowSource {
-    Join {
-        objects: HashMap<ObjectIdPair, Value>,
-        /// TreeObject defining edges
-        // tree_object: TreeObject,
-        root_object: ObjectIdPair,
-        fields: HashMap<String, FieldDefinitionSource>,
-    },
-    Single {
-        root_object: ObjectIdPair,
-        value: Value,
-        fields: HashMap<String, FieldDefinitionSource>,
-    },
 }
 
 impl ObjectBuilderImpl {
